+++ cmd/kobs/plugins/plugins.go
@@ -12,7 +12,6 @@ import (
 	// Import all plugins, which should be used with the kobs instance. By default this are all first party plugins from
 	// the plugins folder.
 	"github.com/kobsio/kobs/plugins/applications"
-	"github.com/kobsio/kobs/plugins/clickhouse"
 	"github.com/kobsio/kobs/plugins/dashboards"
 	"github.com/kobsio/kobs/plugins/elasticsearch"
 	"github.com/kobsio/kobs/plugins/flux"
@@ -21,6 +20,7 @@ import (
 	"github.com/kobsio/kobs/plugins/istio"
 	"github.com/kobsio/kobs/plugins/jaeger"
 	"github.com/kobsio/kobs/plugins/kiali"
+	"github.com/kobsio/kobs/plugins/klogs"
 	"github.com/kobsio/kobs/plugins/markdown"
 	"github.com/kobsio/kobs/plugins/opsgenie"
 	"github.com/kobsio/kobs/plugins/prometheus"
@@ -35,7 +35,6 @@ import (
 // Config holds the configuration for all plugins. We have to add the configuration for all the imported plugins.
 type Config struct {
 	Applications  applications.Config  `json:"applications"`
-	Clickhouse    clickhouse.Config    `json:"clickhouse"`
 	Dashboards    dashboards.Config    `json:"dashboards"`
 	Elasticsearch elasticsearch.Config `json:"elasticsearch"`
 	Flux          flux.Config          `json:"flux"`
@@ -44,6 +43,7 @@ type Config struct {
 	Istio         istio.Config         `json:"istio"`
 	Jaeger        jaeger.Config        `json:"jaeger"`
 	Kiali         kiali.Config         `json:"kiali"`
+	Klogs         klogs.Config         `json:"klogs"`
 	Opsgenie      opsgenie.Config      `json:"opsgenie"`
 	Prometheus    prometheus.Config    `json:"prometheus"`
 	Markdown      markdown.Config      `json:"markdown"`
@@ -84,10 +84,10 @@ func Register(clusters *clusters.Clusters, config Config) chi.Router {
 	dashboardsRouter := dashboards.Register(clusters, router.plugins, config.Dashboards)
 	prometheusRouter, prometheusInstances := prometheus.Register(clusters, router.plugins, config.Prometheus)
 	elasticsearchRouter := elasticsearch.Register(clusters, router.plugins, config.Elasticsearch)
-	clickhouseRouter, clickhouseInstances := clickhouse.Register(clusters, router.plugins, config.Clickhouse)
+	klogsRouter, klogsInstances := klogs.Register(clusters, router.plugins, config.Klogs)
 	jaegerRouter := jaeger.Register(clusters, router.plugins, config.Jaeger)
 	kialiRouter := kiali.Register(clusters, router.plugins, config.Kiali)
-	istioRouter := istio.Register(clusters, router.plugins, config.Istio, prometheusInstances, clickhouseInstances)
+	istioRouter := istio.Register(clusters, router.plugins, config.Istio, prometheusInstances, klogsInstances)
 	grafanaRouter := grafana.Register(clusters, router.plugins, config.Grafana)
 	harborRouter := harbor.Register(clusters, router.plugins, config.Harbor)
 	fluxRouter := flux.Register(clusters, router.plugins, config.Flux)
@@ -105,7 +105,7 @@ func Register(clusters *clusters.Clusters, config Config) chi.Router {
 	router.Mount(dashboards.Route, dashboardsRouter)
 	router.Mount(prometheus.Route, prometheusRouter)
 	router.Mount(elasticsearch.Route, elasticsearchRouter)
-	router.Mount(clickhouse.Route, clickhouseRouter)
+	router.Mount(klogs.Route, klogsRouter)
 	router.Mount(jaeger.Route, jaegerRouter)
 	router.Mount(kiali.Route, kialiRouter)
 	router.Mount(istio.Route, istioRouter)